Create MissingValue-FillWithAverage
Data Preparation For Data Mining
Today, large databases to be accessible, one of the biggest reasons is the emergence and widespread use of data mining. However, specific processes need to be passed in order to allow such data to our study. Some records may be lacking in our database. For example, many people registered in the database, while certain civil status, this information may be missing in some records, so no records can not be entered, This lack of missing data (missing value) can be named. Apart from this, some of the data may include extreme values are meaningless or incorrect data. For example; product codes may have been entered incorrectly. This dirty data (dirty value) can be named. These corrected data during the mining operation will reduce the complexity of computer run time and increase the reliability and quality of the results to be obtained.

Cleaning of Data

Missing and Dirty Data
The missing data that will create techniques that are used to eliminate the problems can be summarized as follows:
1.Missing the record from the database where the data resides (data set), or remove records like this cancel.
If given the number of records lost when the proportion of the total number of records, so it will not affect the accuracy of results is small, this method can be used. Otherwise, this method will bring more harm than benefit.

2. To fill in missing data by hand one by one.
If the database used is small, it is possible to reach this data, and if there's enough time for that, this method can be used. Otherwise, it would be a waste of time.
3. Enter the same information to all of the lost data
4. Missing data the average value of all data to be given
For example, wage data is missing, the part that can be written as the average of the other cost information.
5. With the help of other variables, estimation of missing data. (Learning filling process)
This method allows for missing data to learn the algorithm that must be run. To be filled during the run of the algorithm where the desired information is located of the colon are ignored. Other data outside of the colon are examined, this classification process is performed. Whichever is the highest percentage chance of filling the empty information.

In our project we use the adult data missing data and dirty data were available. It is not possible for us to reach the lost data, we chose the method of filling the average of the methods that have been stated above. For this we have a library by the name of the DataMining.Engine.MissingValue.