Add files via upload
Data Preparation For Data Mining
Today, large databases to be accessible, one of the biggest reasons is the emergence and widespread use of data mining. However, specific processes need to be passed in order to allow such data to our study. Some records may be lacking in our database. For example, many people registered in the database, while certain civil status, this information may be missing in some records, so no records can not be entered, This lack of missing data can be named. Apart from this, some of the data may include extreme values are meaningless or incorrect data. For example; product codes may have been entered incorrectly. This dirty data can be named. These corrected data during the mining operation will reduce the complexity of computer run time and increase the reliability and quality of the results to be obtained.
We chose the method of filling the average of the methods that have been stated above. For this we have a library by the name of the DataMining.Engine.MissingValue.