Add files via upload
Reduction Process Why Is It Necessary?
Reduction process narrows our data. That is, determine specific ranges for each of the available data. Reducing the density and diversity provides a kind of grouping of data. That will occur unless the reduction process is performed for data of a decision tree for each different situation, assesses and expands sideways. A good decision-tree-like structure should not be excessive sideways expansion.

The data values in the data set that we use nested and dispersed. For this reason, we wrote a library will make this process automatic and by the name of DataMining.Engine.Reduction. While doing this process the “3-4-5” rule applied.

3-4-5 Rule Method
The general functioning of this algorithm is as follows:
Inputs: 
One-dimensional data set.
Steps:
1.	Sort the data. From small to large or large to small.
2.	Highest %5 and lowest 5% take off the upper and lower part from the data base.
3.	Find the most obvious step (MOS) between the largest and smallest values of the remaining two data.
4.	Calculate the number of ranges (NOR) between two MOS value.
if ( NOR = 3 or 6 or 7 or 9)
- Split the data into three distinct groups.
- Separating criteria = (New maximum – New minimum) /3
              -      if(NOR = 2 or 4 or 8)
                     - Split the data into four distinct groups..
                     - Separating criteria = (New maximum – New minimum) /4
              -      if(NOR = 1 or 5 or 10)
                     - Split the data into five distinct groups.
                     - Separating criteria = ((New maximum – New minimum) /5

5.	In the second step that are thrown %5 of the top and bottom as the two groups add.