Add files via upload
C 4.5 Algorithm
Missing data are not taken into account when creating this algorithm is a decision tree. That recovery rate is calculated, the data used for other non-missing data. C 4.5 algorithm, anticipating the help of lost data and other data variables used in calculating the rate of recovery. Thus, more sensitive and more can extract meaningful rules that a decision tree can produce. 

In this algorithm, the root node of the decision tree on in the account, while the recovery rate (entropy account) was used.
C 4.5 Algorithm Method

▪	Error estimate for subtree is weighted sum of error estimates for all its leaves
▪	Error estimate for a node (upper bound):
                

▪	If c = 25% then z = 0.69 (from normal distribution)
▪	f  is the error on the training data
▪	N  is the number of instances covered by the leaf